requirements.txt
Flask
python-docx
PyMuPDF
elevenlabs


